requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
chromadb
PyPDF2
requests
huggingface_hub
